Refactored package names
xdzk.cluster
Classes for working with cluster members, in particular with
Containers.

xdzk.curator
Utilities for Curator.

xdzk.server
Server classes, including Admin and Container.
